requirements.txt
Flask==1.1.1
google-cloud-pubsub==1.1.0
google-cloud-logging==1.15.0
google-cloud-firestore==1.6.2
pymongo==3.10.1
dnspython==1.16.0
gunicorn==20.0.4

